Refactor and add sweref99 output
