+++ activefires_pp/tests/test_fires_filtering.py
@@ -18,109 +18,21 @@
 # GNU General Public License for more details.
 
 # You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+# along with this 2program.  If not, see <http://www.gnu.org/licenses/>.
 
 """Test the Fires filtering functionality."""
 
 import pytest
 from unittest.mock import patch
-from unittest import TestCase
-from freezegun import freeze_time
 
 import pandas as pd
-from geojson import FeatureCollection
 import numpy as np
-import io
 from datetime import datetime
 
 from activefires_pp.post_processing import ActiveFiresShapefileFiltering
 from activefires_pp.post_processing import ActiveFiresPostprocessing
 from activefires_pp.post_processing import COL_NAMES
-
-from activefires_pp.post_processing import geojson_feature_collection_from_detections
-
-TEST_ACTIVE_FIRES_FILEPATH = "./AFIMG_j01_d20210414_t1126439_e1128084_b17637_c20210414114130392094_cspp_dev.txt"
-TEST_ACTIVE_FIRES_FILEPATH2 = "./AFIMG_npp_d20230616_t1110054_e1111296_b60284_c20230616112418557033_cspp_dev.txt"
-
-
-TEST_ACTIVE_FIRES_FILE_DATA = """
-# Active Fires I-band EDR
-#
-# source: AFIMG_j01_d20210414_t1126439_e1128084_b17637_c20210414114130392094_cspp_dev.nc
-# version: CSPP Active Fires version: cspp-active-fire-noaa_1.1.0
-#
-# column 1: latitude of fire pixel (degrees)
-# column 2: longitude of fire pixel (degrees)
-# column 3: I04 brightness temperature of fire pixel (K)
-# column 4: Along-scan fire pixel resolution (km)
-# column 5: Along-track fire pixel resolution (km)
-# column 6: detection confidence ([7,8,9]->[lo,med,hi])
-# column 7: fire radiative power (MW)
-#
-# number of fire pixels: 18
-#
-  59.14783859,   37.85886765,  331.58309937,  0.375,  0.375,    8,    4.67825794
-  59.05127335,   28.15227890,  349.83993530,  0.375,  0.375,    8,    7.10289335
-  59.05587006,   28.15146446,  326.76165771,  0.375,  0.375,    8,    7.10289335
-  59.46587372,   29.04332352,  327.60366821,  0.375,  0.375,    8,    5.01662874
-  59.59255981,   28.77226448,  345.88961792,  0.375,  0.375,    8,   13.13724804
-  59.58853149,   28.77531433,  339.56134033,  0.375,  0.375,    8,    8.76600266
-  59.59326553,   28.77456856,  352.21545410,  0.375,  0.375,    8,    8.76600266
-  59.59757233,   28.76391029,  328.43835449,  0.375,  0.375,    8,    5.08633661
-  58.35777283,   12.37761784,  327.17175293,  0.375,  0.375,    8,   17.58141518
-  60.30867004,   25.53105164,  349.98794556,  0.375,  0.375,    8,    6.93412018
-  55.01095581,   -2.28794742,  335.89736938,  0.375,  0.375,    8,    4.39908028
-  59.52483368,   17.16816330,  336.57437134,  0.375,  0.375,    8,   14.13167953
-  55.00822449,   -2.28098702,  344.50894165,  0.375,  0.375,    8,    4.16644764
-  60.13325882,   16.18420029,  329.47689819,  0.375,  0.375,    8,    5.32859230
-  61.30901337,   21.98561668,  341.69180298,  0.375,  0.375,    8,    8.87900448
-  58.29126740,    0.20132475,  331.47875977,  0.375,  0.375,    8,    3.64687872
-  57.42922211,   -3.47403550,  336.02111816,  0.375,  0.375,    8,    8.39092922
-  57.42747116,   -3.47912717,  353.80722046,  0.375,  0.375,    8,   12.13035393
-"""
-
-# Here we have sorted out all detections not passing the filter mask!
-# So, 4 fire detections are left corresponding to what would end up in the geojson files:
-TEST_ACTIVE_FIRES_FILE_DATA2 = """
-# Active Fires I-band EDR
-#
-# source: AFIMG_npp_d20230616_t1110054_e1111296_b60284_c20230616112418557033_cspp_dev.nc
-# version: CSPP Active Fires version: cspp-active-fire-noaa_1.1.0
-#
-# column 1: latitude of fire pixel (degrees)
-# column 2: longitude of fire pixel (degrees)
-# column 3: I04 brightness temperature of fire pixel (K)
-# column 4: Along-scan fire pixel resolution (km)
-# column 5: Along-track fire pixel resolution (km)
-# column 6: detection confidence ([7,8,9]->[lo,med,hi])
-# column 7: fire radiative power (MW)
-#
-# number of fire pixels: 14
-#
-  62.65801239,   17.25905228,  339.66326904,  0.375,  0.375,    8,    2.51202917
-  64.21694183,   17.42074966,  329.65161133,  0.375,  0.375,    8,    3.39806151
-  64.56904602,   16.60095215,  346.52050781,  0.375,  0.375,    8,   20.59289360
-  64.57222748,   16.59840012,  348.72860718,  0.375,  0.375,    8,   20.59289360
-"""
-
-CONFIG_EXAMPLE = {'publish_topic': '/VIIRS/L2/Fires/PP',
-                  'subscribe_topics': 'VIIRS/L2/AFI',
-                  'af_pattern_ibands':
-                  'AFIMG_{platform:s}_d{start_time:%Y%m%d_t%H%M%S%f}_e{end_hour:%H%M%S%f}' +
-                  '_b{orbit:s}_c{processing_time:%Y%m%d%H%M%S%f}_cspp_dev.txt',
-                  'geojson_file_pattern_national': 'AFIMG_{platform:s}_d{start_time:%Y%m%d_t%H%M%S}.geojson',
-                  'geojson_file_pattern_regional': 'AFIMG_{platform:s}_d{start_time:%Y%m%d_t%H%M%S}_' +
-                  '{region_name:s}.geojson',
-                  'regional_shapefiles_format': 'omr_{region_code:s}_Buffer.{ext:s}',
-                  'output_dir': '/path/where/the/filtered/results/will/be/stored',
-                  'filepath_detection_id_cache': '/path/to/the/detection_id/cache',
-                  'timezone': 'Europe/Stockholm'}
-
-OPEN_FSTREAM = io.StringIO(TEST_ACTIVE_FIRES_FILE_DATA)
-
-
-MY_FILE_PATTERN = ("AFIMG_{platform:s}_d{start_time:%Y%m%d_t%H%M%S%f}_e{end_hour:%H%M%S%f}_" +
-                   "b{orbit:s}_c{processing_time:%Y%m%d%H%M%S%f}_cspp_dev.txt")
+from activefires_pp.tests.test_utils import MY_FILE_PATTERN
 
 TEST_REGIONAL_MASK = {}
 TEST_REGIONAL_MASK['Bergslagen (RRB)'] = {'mask': np.array([False, False, False, False, False,
@@ -153,12 +65,11 @@ FAKE_MASK2 = np.array([True, False,  True])
 
 
 @patch('activefires_pp.post_processing._read_data')
-def test_add_start_and_end_time_to_active_fires_data_utc(readdata):
+def test_add_start_and_end_time_to_active_fires_data_utc(readdata, fake_active_fires_file_data):
     """Test adding start and end times to the active fires data."""
-    myfilepath = TEST_ACTIVE_FIRES_FILEPATH
+    open_fstream, myfilepath = fake_active_fires_file_data
+    afdata = pd.read_csv(open_fstream, index_col=None, header=None, comment='#', names=COL_NAMES)
 
-    # fstream = io.StringIO(TEST_ACTIVE_FIRES_FILE_DATA)
-    afdata = pd.read_csv(OPEN_FSTREAM, index_col=None, header=None, comment='#', names=COL_NAMES)
     readdata.return_value = afdata
 
     this = ActiveFiresShapefileFiltering(filepath=myfilepath, timezone='GMT')
@@ -175,12 +86,11 @@ def test_add_start_and_end_time_to_active_fires_data_utc(readdata):
 
 
 @patch('activefires_pp.post_processing._read_data')
-def test_add_start_and_end_time_to_active_fires_data_localtime(readdata):
+def test_add_start_and_end_time_to_active_fires_data_localtime(readdata, fake_active_fires_file_data):
     """Test adding start and end times to the active fires data."""
-    myfilepath = TEST_ACTIVE_FIRES_FILEPATH
+    open_fstream, myfilepath = fake_active_fires_file_data
 
-    fstream = io.StringIO(TEST_ACTIVE_FIRES_FILE_DATA)
-    afdata = pd.read_csv(fstream, index_col=None, header=None, comment='#', names=COL_NAMES)
+    afdata = pd.read_csv(open_fstream, index_col=None, header=None, comment='#', names=COL_NAMES)
     readdata.return_value = afdata
 
     this = ActiveFiresShapefileFiltering(filepath=myfilepath, timezone='Europe/Stockholm')
@@ -231,11 +141,13 @@ def test_add_start_and_end_time_to_active_fires_data_localtime(readdata):
 @patch('socket.gethostname')
 @patch('activefires_pp.post_processing.read_config')
 @patch('activefires_pp.post_processing.ActiveFiresPostprocessing._setup_and_start_communication')
-def test_regional_fires_filtering(setup_comm, get_config, gethostname):
+def test_regional_fires_filtering(setup_comm, get_config, gethostname,
+                                  fake_active_fires_file_data, fake_config_data):
     """Test the regional fires filtering."""
     # FIXME! This test is to big/broad. Need for refactoring!
+    open_fstream, _ = fake_active_fires_file_data
 
-    get_config.return_value = CONFIG_EXAMPLE
+    get_config.return_value = fake_config_data
     gethostname.return_value = "my.host.name"
 
     myconfigfile = "/my/config/file/path"
@@ -245,8 +157,7 @@ def test_regional_fires_filtering(setup_comm, get_config, gethostname):
     afpp = ActiveFiresPostprocessing(myconfigfile, myborders_file, mymask_file)
     afpp._initialize_fire_detection_id()
 
-    fstream = io.StringIO(TEST_ACTIVE_FIRES_FILE_DATA)
-    afdata = pd.read_csv(fstream, index_col=None, header=None, comment='#', names=COL_NAMES)
+    afdata = pd.read_csv(open_fstream, index_col=None, header=None, comment='#', names=COL_NAMES)
 
     starttime = datetime.fromisoformat('2021-04-14 11:26:43.900')
     endtime = datetime.fromisoformat('2021-04-14 11:28:08')
@@ -285,9 +196,13 @@ def test_regional_fires_filtering(setup_comm, get_config, gethostname):
 @patch('activefires_pp.post_processing.read_config')
 @patch('activefires_pp.post_processing.ActiveFiresPostprocessing._setup_and_start_communication')
 @patch('activefires_pp.post_processing.get_global_mask_from_shapefile', side_effect=[FAKE_MASK1, FAKE_MASK2])
-def test_general_national_fires_filtering(get_global_mask, setup_comm, get_config, gethostname):
+def test_general_national_fires_filtering(get_global_mask, setup_comm,
+                                          get_config, gethostname, fake_active_fires_file_data,
+                                          fake_config_data):
     """Test the general/basic national fires filtering."""
-    get_config.return_value = CONFIG_EXAMPLE
+    open_fstream, _ = fake_active_fires_file_data
+
+    get_config.return_value = fake_config_data
     gethostname.return_value = "my.host.name"
 
     myconfigfile = "/my/config/file/path"
@@ -295,9 +210,7 @@ def test_general_national_fires_filtering(get_global_mask, setup_comm, get_confi
     mymask_file = "/my/shape/file/with/polygons/to/filter/out"
 
     afpp = ActiveFiresPostprocessing(myconfigfile, myborders_file, mymask_file)
-
-    fstream = io.StringIO(TEST_ACTIVE_FIRES_FILE_DATA)
-    afdata = pd.read_csv(fstream, index_col=None, header=None, comment='#', names=COL_NAMES)
+    afdata = pd.read_csv(open_fstream, index_col=None, header=None, comment='#', names=COL_NAMES)
 
     # Add metadata to the pandas dataframe:
     fake_metadata = {'platform': 'j01',
@@ -344,9 +257,10 @@ def test_checking_national_borders_shapefile_file_exists(setup_comm, gethostname
 @patch('socket.gethostname')
 @patch('activefires_pp.post_processing.read_config')
 @patch('activefires_pp.post_processing.ActiveFiresPostprocessing._setup_and_start_communication')
-def test_checking_national_borders_shapefile_file_nonexisting(setup_comm, get_config, gethostname):
+def test_checking_national_borders_shapefile_file_nonexisting(setup_comm, get_config, gethostname,
+                                                              fake_config_data):
     """Test the checking of the national borders shapefile - borders shapefile does not exist."""
-    get_config.return_value = CONFIG_EXAMPLE
+    get_config.return_value = fake_config_data
     gethostname.return_value = "my.host.name"
 
     myconfigfile = "/my/config/file/path"
@@ -359,76 +273,3 @@ def test_checking_national_borders_shapefile_file_nonexisting(setup_comm, get_co
 
     expected = "Shape file does not exist! Filename = /my/shape/file/with/country/borders"
     assert str(exec_info.value) == expected
-
-
-@freeze_time('2023-06-16 11:24:00')
-@patch('socket.gethostname')
-@patch('activefires_pp.post_processing.read_config')
-@patch('activefires_pp.post_processing.ActiveFiresPostprocessing._setup_and_start_communication')
-@patch('activefires_pp.post_processing._read_data')
-def test_get_feature_collection_from_firedata(readdata, setup_comm,
-                                              get_config, gethostname):
-    """Test get the Geojson Feature Collection from fire detection."""
-    get_config.return_value = CONFIG_EXAMPLE
-    gethostname.return_value = "my.host.name"
-
-    myconfigfile = "/my/config/file/path"
-    myborders_file = "/my/shape/file/with/country/borders"
-    mymask_file = "/my/shape/file/with/polygons/to/filter/out"
-
-    afpp = ActiveFiresPostprocessing(myconfigfile, myborders_file, mymask_file)
-    afpp._initialize_fire_detection_id()
-
-    myfilepath = TEST_ACTIVE_FIRES_FILEPATH2
-
-    fstream = io.StringIO(TEST_ACTIVE_FIRES_FILE_DATA2)
-    afdata = pd.read_csv(fstream, index_col=None, header=None, comment='#', names=COL_NAMES)
-    readdata.return_value = afdata
-
-    this = ActiveFiresShapefileFiltering(filepath=myfilepath, timezone='GMT')
-    with patch('os.path.exists') as mypatch:
-        mypatch.return_value = True
-        afdata = this.get_af_data(filepattern=MY_FILE_PATTERN, localtime=False)
-
-    afdata = afpp.add_unique_day_id(afdata)
-
-    result = geojson_feature_collection_from_detections(afdata, platform_name='Suomi-NPP')
-
-    # NB! The time of the afdata is here still in UTC!
-    expected = FeatureCollection([{"geometry": {"coordinates": [17.259052, 62.658012],
-                                                "type": "Point"},
-                                   "properties": {"confidence": 8,
-                                                  "observation_time": "2023-06-16T11:10:47.200000",
-                                                  "platform_name": "Suomi-NPP",
-                                                  "id": '20230616-1',
-                                                  "power": 2.51202917, "tb": 339.66326904},
-                                   "type": "Feature"},
-                                  {"geometry": {"coordinates": [17.42075, 64.216942],
-                                                "type": "Point"},
-                                   "properties": {"confidence": 8,
-                                                  "observation_time": "2023-06-16T11:10:47.200000",
-                                                  "platform_name": "Suomi-NPP",
-                                                  "id": '20230616-2',
-                                                  "power": 3.39806151,
-                                                  "tb": 329.65161133},
-                                   "type": "Feature"},
-                                  {"geometry": {"coordinates": [16.600952, 64.569046],
-                                                "type": "Point"},
-                                   "properties": {"confidence": 8,
-                                                  "observation_time": "2023-06-16T11:10:47.200000",
-                                                  "platform_name": "Suomi-NPP",
-                                                  "id": '20230616-3',
-                                                  "power": 20.5928936,
-                                                  "tb": 346.52050781},
-                                   "type": "Feature"},
-                                  {"geometry": {"coordinates": [16.5984, 64.572227],
-                                                "type": "Point"},
-                                   "properties": {"confidence": 8,
-                                                  "observation_time": "2023-06-16T11:10:47.200000",
-                                                  "platform_name": "Suomi-NPP",
-                                                  "id": '20230616-4',
-                                                  "power": 20.5928936,
-                                                  "tb": 348.72860718},
-                                   "type": "Feature"}])
-
-    TestCase().assertDictEqual(result, expected)
