+++ activefires_pp/tests/test_geojson_utils.py
@@ -22,14 +22,31 @@
 
 """Test the Geojson utillities."""
 
-from activefires_pp.geojson_utils import read_geojson_data
-from activefires_pp.geojson_utils import get_geojson_files_in_observation_time_order
-from activefires_pp.geojson_utils import store_geojson_alarm
-from trollsift import Parser
+from unittest.mock import patch
+from unittest import TestCase
 
+from freezegun import freeze_time
 from datetime import datetime
 import pytest
 import logging
+import pandas as pd
+import numpy as np
+from geojson import FeatureCollection
+
+from trollsift import Parser
+
+from activefires_pp.geojson_utils import store_geojson
+from activefires_pp.geojson_utils import geojson_feature_collection_from_detections
+from activefires_pp.geojson_utils import read_geojson_data
+from activefires_pp.geojson_utils import get_geojson_files_in_observation_time_order
+from activefires_pp.geojson_utils import store_geojson_alarm
+from activefires_pp.geojson_utils import map_coordinates_in_feature_collection
+
+from activefires_pp.post_processing import ActiveFiresShapefileFiltering
+from activefires_pp.post_processing import ActiveFiresPostprocessing
+from activefires_pp.post_processing import COL_NAMES
+from activefires_pp.tests.test_utils import MY_FILE_PATTERN
+
 
 TEST_GEOJSON_FILE_CONTENT = """{"type": "FeatureCollection", "features":
 [{"type": "Feature", "geometry": {"type": "Point", "coordinates": [23.562864, 67.341919]},
@@ -174,3 +191,152 @@ def test_store_geojson_alarm(fake_past_detections_dir):
     assert json_test_data['features']['properties']['power'] == 2.23312426
     assert json_test_data['features']['properties']['related_detection'] is False
     assert json_test_data['features']['properties']['tb'] == 310.37322998
+
+
+def test_store_geojson_file_sweref99_coordinates(tmp_path):
+    """Test store Geojson file to disk."""
+    feature_collection = {"features": [{"geometry": {"coordinates": [2804994.83249444, 871459.9503322293],
+                                                     "type": "Point"},
+                                        "properties": {"confidence": 8,
+                                                       "id": "20230616-1",
+                                                       "observation_time": "2023-06-16T11:10:47.200000",
+                                                       "platform_name": "Suomi-NPP",
+                                                       "power": 2.51202917,
+                                                       "tb": 339.66326904},
+                                        "type": "Feature"}, {
+                                            "geometry": {"coordinates": [2654228.542928629, 832840.571493448],
+                                                         "type": "Point"},
+                                            "properties": {"confidence": 8,
+                                                           "id": "20230616-2",
+                                                           "observation_time": "2023-06-16T11:10:47.200000",
+                                                           "platform_name": "Suomi-NPP",
+                                                           "power": 3.39806151,
+                                                           "tb": 329.65161133},
+                                            "type": "Feature"}],
+                          "type": "FeatureCollection"}
+
+    output_path = tmp_path / 'test_fire_detections.geojson'
+    store_geojson(output_path, feature_collection)
+    assert output_path.exists()
+
+    json_test_data = read_geojson_data(output_path)
+    np.testing.assert_almost_equal(json_test_data['features'][0]['geometry']['coordinates'],
+                                   np.array([2804994.83249444, 871459.9503322293]), decimal=6)
+    np.testing.assert_almost_equal(json_test_data['features'][1]['geometry']['coordinates'],
+                                   np.array([2654228.542928629, 832840.571493448]), decimal=6)
+
+
+@freeze_time('2023-06-16 11:24:00')
+@patch('socket.gethostname')
+@patch('activefires_pp.post_processing.read_config')
+@patch('activefires_pp.post_processing.ActiveFiresPostprocessing._setup_and_start_communication')
+@patch('activefires_pp.post_processing._read_data')
+def test_get_feature_collection_from_firedata(readdata, setup_comm,
+                                              get_config, gethostname,
+                                              fake_active_fires_file_data2, fake_config_data):
+    """Test get the Geojson Feature Collection from fire detection."""
+    open_fstream, myfilepath = fake_active_fires_file_data2
+    get_config.return_value = fake_config_data
+    gethostname.return_value = "my.host.name"
+
+    myconfigfile = "/my/config/file/path"
+    myborders_file = "/my/shape/file/with/country/borders"
+    mymask_file = "/my/shape/file/with/polygons/to/filter/out"
+
+    afpp = ActiveFiresPostprocessing(myconfigfile, myborders_file, mymask_file)
+    afpp._initialize_fire_detection_id()
+
+    afdata = pd.read_csv(open_fstream, index_col=None, header=None, comment='#', names=COL_NAMES)
+    readdata.return_value = afdata
+
+    this = ActiveFiresShapefileFiltering(filepath=myfilepath, timezone='GMT')
+    with patch('os.path.exists') as mypatch:
+        mypatch.return_value = True
+        afdata = this.get_af_data(filepattern=MY_FILE_PATTERN, localtime=False)
+
+    afdata = afpp.add_unique_day_id(afdata)
+
+    result = geojson_feature_collection_from_detections(afdata, platform_name='Suomi-NPP')
+
+    # NB! The time of the afdata is here still in UTC!
+    expected = FeatureCollection([{"geometry": {"coordinates": [17.259052, 62.658012],
+                                                "type": "Point"},
+                                   "properties": {"confidence": 8,
+                                                  "observation_time": "2023-06-16T11:10:47.200000",
+                                                  "platform_name": "Suomi-NPP",
+                                                  "id": '20230616-1',
+                                                  "power": 2.51202917, "tb": 339.66326904},
+                                   "type": "Feature"},
+                                  {"geometry": {"coordinates": [17.42075, 64.216942],
+                                                "type": "Point"},
+                                   "properties": {"confidence": 8,
+                                                  "observation_time": "2023-06-16T11:10:47.200000",
+                                                  "platform_name": "Suomi-NPP",
+                                                  "id": '20230616-2',
+                                                  "power": 3.39806151,
+                                                  "tb": 329.65161133},
+                                   "type": "Feature"},
+                                  {"geometry": {"coordinates": [16.600952, 64.569046],
+                                                "type": "Point"},
+                                   "properties": {"confidence": 8,
+                                                  "observation_time": "2023-06-16T11:10:47.200000",
+                                                  "platform_name": "Suomi-NPP",
+                                                  "id": '20230616-3',
+                                                  "power": 20.5928936,
+                                                  "tb": 346.52050781},
+                                   "type": "Feature"},
+                                  {"geometry": {"coordinates": [16.5984, 64.572227],
+                                                "type": "Point"},
+                                   "properties": {"confidence": 8,
+                                                  "observation_time": "2023-06-16T11:10:47.200000",
+                                                  "platform_name": "Suomi-NPP",
+                                                  "id": '20230616-4',
+                                                  "power": 20.5928936,
+                                                  "tb": 348.72860718},
+                                   "type": "Feature"}])
+
+    TestCase().assertDictEqual(result, expected)
+
+
+def test_map_coordinates_in_feature_collection_sweref99():
+    """Test mapping the coordinates to SWEREF99."""
+    fc_in = FeatureCollection([{"geometry": {"coordinates": [17.259052, 62.658012],
+                                             "type": "Point"},
+                                "properties": {"confidence": 8,
+                                               "observation_time": "2023-06-16T11:10:47.200000",
+                                               "platform_name": "Suomi-NPP",
+                                               "id": '20230616-1',
+                                               "power": 2.51202917, "tb": 339.66326904},
+                                "type": "Feature"},
+                               {"geometry": {"coordinates": [17.42075, 64.216942],
+                                             "type": "Point"},
+                                "properties": {"confidence": 8,
+                                               "observation_time": "2023-06-16T11:10:47.200000",
+                                               "platform_name": "Suomi-NPP",
+                                               "id": '20230616-2',
+                                               "power": 3.39806151,
+                                               "tb": 329.65161133},
+                                "type": "Feature"}])
+
+    expected = FeatureCollection([{"geometry": {"coordinates": [2804994.83249444, 871459.9503322293],
+                                                "type": "Point"},
+                                   "properties": {"confidence": 8,
+                                                  "id": "20230616-1",
+                                                  "observation_time": "2023-06-16T11:10:47.200000",
+                                                  "platform_name": "Suomi-NPP",
+                                                  "power": 2.51202917,
+                                                  "tb": 339.66326904},
+                                   "type": "Feature"},
+                                  {"geometry": {"coordinates": [2654228.542928629, 832840.571493448],
+                                                "type": "Point"},
+                                   "properties": {"confidence": 8,
+                                                  "id": "20230616-2",
+                                                  "observation_time": "2023-06-16T11:10:47.200000",
+                                                  "platform_name": "Suomi-NPP",
+                                                  "power": 3.39806151,
+                                                  "tb": 329.65161133},
+                                   "type": "Feature"}])
+
+    fc_out = map_coordinates_in_feature_collection(fc_in, '4976')
+
+    TestCase().assertDictEqual(fc_out, expected)
